Keep session alive even if websocket disconnect
The current behavior is to create a new "game" on websocket
reconnection. This is not ideal when the user locks its screen, as it
creates an unwanted disconnection.

From now on, the session will persists through disconnections, as we
store a token (per tab, in session), to keep track of the current
sessions id and send it back to the backend to continue playing.